fix(strings): renamed code for strings
