+++ src/strings.js
